Tweaks
Added Item ready cycle action for rules

With this action, a rule can be setup to trigger on cron run, that will call this action. 

This action will pull up the time that the last item read was requested, and create a request for items changed after that time.
This will create a request that will send up the modified data in a cycle